feat(web): rebuild schema-driven Constituter
