+++ tests/unit/web-configurator.test.ts
@@ -2,9 +2,11 @@ import { describe, expect, it } from 'vitest';
 import { normalizeConfig, type ConfigInput } from '@hipster-stack/core';
 import {
   createCliCommand,
+  createShareUrl,
   defaultConfiguratorRecipe,
   deserializeRecipe,
   resolveConfiguratorRecipe,
+  selectProductPreset,
   serializeRecipe,
   setCapability,
 } from '../../apps/web/lib/configurator.js';
@@ -30,6 +32,38 @@ describe('web configurator recipe', () => {
   it('provides the canonical Hipster Stack package command', () => {
     expect(createCliCommand(defaultConfiguratorRecipe)).toBe(
       'pnpm dlx hipster-stack@latest my-saas --config hipsterstack.json --yes',
+    );
+  });
+
+  it('applies a preset through shared resolution without retaining overrides', () => {
+    const customized = setCapability(
+      defaultConfiguratorRecipe,
+      'stripeConnect',
+      true,
+    );
+    const selected = selectProductPreset(customized, 'client-portal');
+
+    expect(selected.modules).toEqual({});
+    expect(resolveConfiguratorRecipe(selected).summary.preset.id).toBe(
+      'client-portal',
+    );
+    expect(resolveConfiguratorRecipe(selected).recipe.modules.onboarding).toBe(
+      true,
+    );
+  });
+
+  it('creates a portable share URL that hydrates the same recipe', () => {
+    const shared = createShareUrl(
+      defaultConfiguratorRecipe,
+      'https://hipster.example/configure?source=test',
+    );
+    const url = new URL(shared);
+    const encoded = url.searchParams.get('recipe');
+
+    expect(url.searchParams.get('source')).toBe('test');
+    expect(encoded).not.toBeNull();
+    expect(serializeRecipe(deserializeRecipe(encoded!))).toBe(
+      serializeRecipe(defaultConfiguratorRecipe),
     );
   });
 });
